Refactor new position to components and views
