+++ frontend/src/routes.js
@@ -2,7 +2,7 @@ import Positions from "./modules/positions/components/index";
 import Applicants from "./modules/applicants_by_course/components/index";
 import RootRedirect from "./modules/auth/components/RootRedirect";
 import Screen from "./modules/cp/components/index";
-import NewPosition from "./modules/positions/components/NewPosition";
+import NewPosition from "./views/new-position-form/fillablePosition";
 import AllAssigned from "./modules/all_applicants/components/AllAssigned";
 import AllUnassigned from "./modules/all_applicants/components/AllUnassigned";
 import Summary from "./modules/tapp_summary/components/Summary";
