+++ frontend/src/views/new-position-form/constants.js
@@ -1,0 +1,3 @@
+export const CREATE_NEW_POSITION_SUCCESS = "CREATE_NEW_POSITION_SUCCESS";
+export const IMPORT_NEW_POSITION_SUCCESS = "IMPORT_NEW_POSITION_SUCCESS";
+export const IMPORT_NEW_POSITION_RESULT = "IMPORT_NEW_POSITION_RESULT";
