--- frontend/src/views/new_position_form/constants.js
@@ -1,8 +1,0 @@
-export const FETCH_POSITIONS_SUCCESS = "FETCH_POSITIONS_SUCCESS";
-export const SAVE_POSITION_SUCCESS = "SAVE_POSITION_SUCCESS";
-export const OPEN_EDIT_POSITION_MODAL = "OPEN_EDIT_POSITION_MODAL";
-export const CLOSE_EDIT_POSITION_MODAL = "CLOSE_EDIT_POSITION_MODAL";
-export const DELETE_POSITION_SUCCESS = "DELETE_POSITION_SUCCESS";
-export const CREATE_NEW_POSITION_SUCCESS = "CREATE_NEW_POSITION_SUCCESS";
-export const IMPORT_NEW_POSITION_SUCCESS = "IMPORT_NEW_POSITION_SUCCESS";
-export const IMPORT_NEW_POSITION_RESULT = "IMPORT_NEW_POSITION_RESULT";
